requirements.txt
pyserial
numpy
pandas
scikit-learn
tensorflow
flask
flask-cors
matplotlib
requests
pykrige
joblib
